Enable running cache inconsistency detection by default

+++ staging/src/k8s.io/apiserver/pkg/storage/cacher/watch_cache.go
@@ -936,9 +936,9 @@ func (w *watchCache) getIntervalFromStoreLocked(key string, matchesSingle bool) 
 	return ci, nil
 }
 
-func (w *watchCache) Compact(rev int64) {
+func (w *watchCache) Compact(rev uint64) {
 	if w.snapshots == nil {
 		return
 	}
-	w.snapshots.RemoveLess(uint64(rev))
+	w.snapshots.RemoveLess(rev)
 }
